1.0
Solution for Project Euler's Problem 14:
https://projecteuler.net/problem=14

This isn't the original solution I used, which used a brute force approach on each path; this one, however, is considerably faster and can actually finish within a day or so. Note, however, that it does break CMD's default recursion stack, so, to run this script, CMD.exe has to be modified with editbin to support a larger recursion stack. Instructions are provided in the comments at the start of the script.